runtime: split syscall functions for darwin changes for 1.26
